Migrated loki syntax and labels logic to useLokiSyntax hook
